+++ c++/hittable.hpp
@@ -9,10 +9,10 @@ struct hit_record {
     point3 p;
     vec3 normal;
     shared_ptr<material> mat_ptr;
-    double t;
+    float t;
 };
 
 class hittable {
     public:
-        virtual bool hit(const ray& r, double t_min, double t_max, hit_record& rec) const = 0;
+        virtual bool hit(const ray& r, float t_min, float t_max, hit_record& rec) const = 0;
 };
